pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vism_lib"
version = "0.1.1"
authors = [
  { name="atrivisc", email="atrivisc.shawl130@passmail.com" },
]
description = "Library for vism CA and ACME"
readme = "README.md"
requires-python = ">=3.14"
license = { file = "LICENSE" }
keywords = ["vism"]
dependencies = [
    "python-pkcs11>=0.9.3",
    "yaml-pydantic-dataclass>=0.1.1",
    "pydantic>=2.12.5",
    "starlette>=0.52.1"

]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/atrivisc/vism-lib"
"Repository" = "https://github.com/atrivisc/vism-lib.git"
"Bug Tracker" = "https://github.com/atrivisc/vism-lib/issues"

[tool.setuptools.packages.find]
where = ["src"]
